+++ OpenAI.SDK/ObjectModels/ResponseModels/ImageResponseModel/ImageCreateResponse.cs
@@ -7,7 +7,7 @@ public record ImageCreateResponse : BaseResponse, IOpenAiModels.ICreatedAt
 {
     [JsonPropertyName("data")] public List<ImageDataResult> Results { get; set; }
 
-    [JsonPropertyName("created")] public int CreatedAt { get; set; }
+    [JsonPropertyName("created")] public long CreatedAt { get; set; }
 
     public record ImageDataResult
     {
