+++ OpenAI.SDK/ObjectModels/SharedModels/FileResponse.cs
@@ -10,6 +10,6 @@ public record FileResponse : BaseResponse, IOpenAiModels.IId, IOpenAiModels.ICre
     public UploadFilePurposes.UploadFilePurpose PurposeEnum => UploadFilePurposes.ToEnum(Purpose);
     [JsonPropertyName("purpose")] public string Purpose { get; set; }
     [JsonPropertyName("status")] public string Status { get; set; }
-    [JsonPropertyName("created_at")] public int CreatedAt { get; set; }
+    [JsonPropertyName("created_at")] public long CreatedAt { get; set; }
     [JsonPropertyName("id")] public string Id { get; set; }
 }
